+++ src/cmdliner_eval.ml
@@ -149,13 +149,24 @@ let env_default v = try Some (Sys.getenv v) with Not_found -> None
 let remove_exec argv =
   try List.tl (Array.to_list argv) with Failure _ -> invalid_arg err_argv
 
+let do_deprecated_msgs err_ppf cl ei =
+  let cmd = Cmdliner_info.Eval.cmd ei in
+  let msgs = Cmdliner_cline.deprecated_msgs cl in
+  let msgs = match Cmdliner_info.Cmd.deprecated cmd with
+  | None -> msgs
+  | Some msg ->
+      let name = Cmdliner_base.quote (Cmdliner_info.Cmd.name cmd) in
+      String.concat "" ("command " :: name :: ": " :: msg :: []) :: msgs
+  in
+  Cmdliner_msg.pp_err err_ppf ei ~err:(String.concat "\n" msgs)
+
 let eval_value
     ?help:(help_ppf = Format.std_formatter)
     ?err:(err_ppf = Format.err_formatter)
     ?(catch = true) ?(env = env_default) ?(argv = Sys.argv) cmd
   =
-  let args, f, i, parents, res = find_term (remove_exec argv) cmd in
-  let ei = Cmdliner_info.Eval.v ~cmd:i ~parents ~env in
+  let args, f, cmd, parents, res = find_term (remove_exec argv) cmd in
+  let ei = Cmdliner_info.Eval.v ~cmd ~parents ~env in
   let help, version, ei = add_stdopts ei in
   let term_args = Cmdliner_info.Cmd.args @@ Cmdliner_info.Eval.cmd ei in
   let res = match res with
@@ -178,10 +189,7 @@ let eval_value
           match try_eval_stdopts ~catch ei cl help version with
           | Some e -> e
           | None ->
-              begin match Cmdliner_info.(Cmd.deprecated (Eval.cmd ei)) with
-              | None -> ()
-              | Some err -> Cmdliner_msg.pp_err err_ppf ei ~err
-              end;
+              do_deprecated_msgs err_ppf cl ei;
               run_parser ~catch ei cl f
   in
   do_result help_ppf err_ppf ei res